% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouping-functions.R
\name{get_cluster_members}
\alias{get_cluster_members}
\title{Finds all leaves of a cluster in a tree data structure.}
\usage{
get_cluster_members(hc, cluster_index)
}
\arguments{
\item{hc}{is a hclust object.}

\item{cluster_index}{is The index of the cluster in the hclust object.}
}
\value{
\code{integers} representing all leafs (retention time indexes) of a given
node (and sub-nodes) in the tree
}
\description{
Auxillary function for \code{groupSimilarityMatrixTree()} that searches through a
complete clustering tree for branches with internal distances less
than maxDiff.
}
\details{
The algorithm recursively seeks the tree for the members of a cluster.
}
\seealso{
Other grouping operations: 
\code{\link{groupClosest}()},
\code{\link{groupConsecutive}()},
\code{\link{groupSimilarityMatrix}()},
\code{\link{groupSimilarityMatrixTree}()}
}
\author{
Johan Lassen
}
\concept{grouping operations}
