% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbundanceSimilarityParam.R
\docType{class}
\name{groupFeatures-similar-abundance}
\alias{groupFeatures-similar-abundance}
\alias{AbundanceSimilarityParam}
\title{Group features based on abundance similarities across samples}
\usage{
AbundanceSimilarityParam(
  threshold = 0.9,
  simFun = cor,
  groupFun = groupSimilarityMatrix,
  subset = integer(),
  transform = log2,
  ...
)
}
\arguments{
\item{threshold}{\code{numeric(1)} defining the (similarity) threshold to be used
for the feature grouping. This parameter is passed to the \code{groupFun}
function.}

\item{simFun}{\code{function} to be used to calculate the (pairwise)
similarities. Defaults to \code{simFun = cor} in which case also
\code{use = "pairwise.complete.obs"} is automatically passed along to the
function.}

\item{groupFun}{\code{function} to group features based on the calculated
similarity matrix. Defaults to \code{groupFun = groupSimilarityMatrix}. See
\code{\link[=groupSimilarityMatrix]{groupSimilarityMatrix()}} for details.}

\item{subset}{\code{integer} or \code{logical} defining a subset of samples (at least
2) on which the similarity calculation should be performed. By default
the calculation is performed on all samples.}

\item{transform}{\code{function} to be used to transform feature abundances prior
to the similarity calculation. Defaults to \code{transform = log2}. To use
feature values \emph{as is} use \code{transform = identity}.}

\item{...}{for \code{AbundanceSimilarityParam}: optional parameters to be passed
along to \code{simFun} and \code{groupFun}. For \code{groupFeatures}: optional
parameters for the extraction/definition of the feature values from
\code{object}.}
}
\description{
Group features based on similar abundances (i.e. \emph{feature values}) across
samples. Parameter \code{subset} allows to define a sub set of samples on which
the similarity calculation should be performed. It might for example be
better to exclude QC samples from the analysis because feature values  are
supposed to be constant in these samples.

The function first calculates a nxn similarity matrix with n being the
number of features and subsequently groups features for which the similarity
is higher than the user provided threshold. Parameter \code{simFun} allows
to specify the function to calculate the pairwise similarities on the feature
values (eventually transformed by the function specified with parameter
\code{transform}). \code{simFun} defaults to \code{cor} but any function that
returns a (symmetric) numeric similarity matrix can be used. Parameter
\code{groupFun} allows to specify the function to group the features based on the
similarity function. It defaults to \code{groupSimilarityMatrix}. See
\code{\link[=groupSimilarityMatrix]{groupSimilarityMatrix()}} for details.

Additional settings for the \code{groupFun} and \code{simFun} functions can be passed
to the \strong{parameter object}  with the \code{...} in the \code{AbundanceSimilarityParam}
constructor function. Other additional parameters specific for the type
of \code{object} can be passed \emph{via} \code{...} in the \code{groupFeatures} call.
}
\seealso{
\code{\link[=groupFeatures]{groupFeatures()}} for the general concept of feature grouping.

\code{\link[=featureGroups]{featureGroups()}} for the function to extract defined feature
groups from a \code{SummarizedExperiment}.

Other feature grouping methods: 
\code{\link{groupFeatures-similar-rtime}}
}
\author{
Johannes Rainer
}
\concept{feature grouping methods}
