% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimilarRtimeParam.R
\name{groupFeatures-similar-rtime}
\alias{groupFeatures-similar-rtime}
\alias{SimilarRtimeParam}
\alias{groupFeatures,numeric,SimilarRtimeParam-method}
\title{Group features based on approximate retention times}
\usage{
SimilarRtimeParam(diffRt = 1, groupFun = groupClosest)

\S4method{groupFeatures}{numeric,SimilarRtimeParam}(object, param, ...)
}
\arguments{
\item{diffRt}{\code{numeric(1)} defining the retention time window within which
features should be grouped. All features with a rtime difference
smaller or equal than \code{diffRt} are grouped.}

\item{groupFun}{\code{function} that can be used to group values. Defaults to
\code{groupFun = groupClosest}. See description for details and alternatives.}

\item{object}{input object that provides the retention times that should be
grouped. The \code{MsFeatures} package defines a method for \code{object} being
either a \code{numeric} or a \code{SummarizedExperiment}.}

\item{param}{\code{SimilarRtimeParam} object with the settings for the method.}

\item{...}{additional parameters passed to the \code{groupFun} function.}
}
\value{
Depending on parameter \code{object}:
\itemize{
\item for \code{object} being a \code{numeric}: returns a \code{factor} defining the feature
groups.
}
}
\description{
Group features based on similar retention time. This method is supposed to be
used as an initial \emph{crude} grouping of LC-MS features based on the median
retention time of all their chromatographic peaks. All features with a
difference in their retention time which is \code{<=} parameter \code{diffRt} of the
parameter object are grouped together. If a column \code{"feature_group"} is
found in
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment::colData()}} this is further sub-grouped by this method.

Parameter \code{groupFun} allows to specify the function that should be used for
the actual grouping. Two possible choices are:
\itemize{
\item \code{groupFun = groupClosest} (the default): the \code{\link[=groupClosest]{groupClosest()}} function
groups values together if their difference is smaller than \code{diffRt}.
If the difference of a feature to more than one group is smaller \code{diffRt}
it is assigned to the group to which its retention time is closest (most
similar) to the mean retention time of that group. This leads to smaller
group sizes. See \code{\link[=groupClosest]{groupClosest()}} for details and examples.
\item \code{groupFun = MsCoreUtils::group}: this function consecutively groups
elements together if their difference in retention time is smaller than
\code{diffRt}. If two features are grouped into one group, also all other
features with a retention time within the defined window to any of the two
features are also included into the feature group. This grouping is
recursively expanded which can lead, depending on \code{diffRt}, to very large
feature groups spanning a large retention time window. See
\code{\link[MsCoreUtils:group]{MsCoreUtils::group()}} for details.
}

Other grouping functions might be added in future. Alternatively it is also
possible to provide a custom function for the grouping operation.
}
\examples{

## Simple grouping of a numeric vector.
##
## Define a numeric vector that could represent retention times of features
x <- c(2, 3, 4, 5, 10, 11, 12, 14, 15)

## Group the values using a `group` function. This will create larger
## groups.
groupFeatures(x, param = SimilarRtimeParam(2, MsCoreUtils::group))

## Group values using the default `groupClosest` function. This creates
## smaller groups in which all elements have a difference smaller than the
## defined `diffRt` with each other.
groupFeatures(x, param = SimilarRtimeParam(2, groupClosest))
}
\seealso{
\code{\link[=groupFeatures]{groupFeatures()}} for the general concept of feature grouping.
}
\author{
Johannes Rainer
}
\concept{feature grouping methods}
