% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R
\name{se}
\alias{se}
\title{Quantified LC-MS preprocessing result test data}
\description{
The \code{se} variable is a \code{\link[=SummarizedExperiment]{SummarizedExperiment()}} object representing the
results from a \code{xcms}-based pre-processing of an LC-MS untargeted
metabolomics data set. The raw data files are provided in the \code{faahKO}
package. The pre-processing of this data set is described in detail in
the \emph{xcms} vignette of the \code{xcms} package. This object was created from the
\code{XCMSnExp} result object with the \code{quantify} method.
}
\examples{

## Load the data
data(se)

library(SummarizedExperiment)

## Access row (feature) data
rowData(se)
}
