% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimilarRtimeParam.R
\name{groupFeatures-similar-rtime}
\alias{groupFeatures-similar-rtime}
\alias{SimilarRtimeParam}
\alias{groupFeatures,numeric,SimilarRtimeParam-method}
\alias{groupFeatures,XCMSnExp,SimilarRtimeParam-method}
\alias{groupFeatures,SummarizedExperiment,SimilarRtimeParam-method}
\title{Group features based on similar retention times}
\usage{
SimilarRtimeParam(diffRt = 1, method = c("groupClosest", "greedy"))

\S4method{groupFeatures}{numeric,SimilarRtimeParam}(object, param)

\S4method{groupFeatures}{XCMSnExp,SimilarRtimeParam}(object, param, msLevel = 1L)

\S4method{groupFeatures}{SummarizedExperiment,SimilarRtimeParam}(object, param, column = "rtime")
}
\arguments{
\item{diffRt}{\code{numeric(1)} defining the retention time window within which
features should be grouped. All features with a rtime difference
smaller or equal than \code{diffRt} are grouped.}

\item{method}{\code{character(1)} defining which grouping approach should be
taken. Allowed values are \code{method = "groupClosest"} (the default) and
\code{method = "greedy"}. See description for details.}

\item{object}{can be a
- \code{\link[=XCMSnExp]{XCMSnExp()}} object containing correspondence results (i.e. feature
definitions).
- \code{numeric} with the retention times.
- \code{\link[=SummarizedExperiment]{SummarizedExperiment()}} object were each row is considered to
represent one \emph{feature}. One column in its \code{rowData} is supposed to
contain the retention times.}

\item{param}{\code{SimilarRtimeParam} object with the settings for the method.}

\item{msLevel}{for \verb{groupFeatures,XCMSnExp}: \code{integer(1)} defining the MS
level on which the features should be grouped.}

\item{column}{for \verb{groupFeatures,SummarizedExperiment}: \code{character(1)}
defining the column in the object's \code{rowData} containing the retention
times on which the features (rows) should be grouped. Defaults to
\code{column = "rtime"}.}
}
\value{
depending on the input \code{object}:
\itemize{
\item \code{XCMSnExp}: the input object with feature groups added in column
\code{"feature_group"} of its \code{featureDefinitions} data frame.
\item \code{SummarizedExperiment}: the input object with feature groups added in
column \code{"feature_group"} in the object's \code{rowData}.
\item \code{numeric}: an \code{integer} with the group assignment of each value.
}
}
\description{
Group features based on similar retention time. This method is supposed to be
used as an initial \emph{crude} grouping of co-eluting features based on their
retention. All features with a difference in their retention time
\code{<=} parameter \code{diffRt} of the \code{SimilarRtimeParam} object
are grouped together. If this method is called on an \code{XCMSnExp} object or an
\code{SummarizedExperiment} and a column \code{"feature_group"} is present in
its \code{\link[xcms:XCMSnExp-class]{xcms::featureDefinitions()}} or \link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment::rowData} data
frame, respectively, these feature groups are further sub-grouped by this
method.

Two different grouping methods are available:
\itemize{
\item \code{method = "greedy"}: this approach consecutively groups elements together
if their difference in retention time is smaller than \code{diffRt}. If two
features are grouped into one group, also all other features with a
retention time within the defined window to any of the two features are
also included into the feature group. This grouping is recursively
expanded which can lead, depending on \code{diffRt} to very large feature
groups spanning a large retention time window.
\item \code{method = "groupClosest"}: this approach uses the \code{\link[=groupClosest]{groupClosest()}} function
that groups values together if their difference is smaller than \code{diffRt}.
If the difference of a feature to more than one group is smaller \code{diffRt}
it is assigned to the group to which its retention time is closest (most
similar) to the mean retention time of that group. This leads to smaller
group sizes. See \code{\link[=groupClosest]{groupClosest()}} for details and examples.
}
}
\examples{

### --------------------------------------------- ###
## Grouping a simple numeric vector

x <- c(3, 5, 3, 4, 12, 4, 13, 11)
groupFeatures(x, SimilarRtimeParam(1))


### --------------------------------------------- ###
## Feature grouping on a result object from xcms

library(xcms)
## Loading the example data
data("xdata")
## Update the path to the files for the local system
xcms::dirname(xdata) <- file.path(system.file("cdf", package = "faahKO"),
                               rep(c("KO/", "WT/"), each = 4))

## The feature definitions
featureDefinitions(xdata)

res <- groupFeatures(xdata, SimilarRtimeParam(25))

## The feature groups
table(featureDefinitions(res)$feature_group)
}
\author{
Johannes Rainer
}
\concept{feature grouping methods}
