% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidden_aliases.R, R/AllGenerics.R
\docType{methods}
\name{hidden_aliases}
\alias{hidden_aliases}
\alias{SimilarRtimeParam-class}
\alias{groupFeatures}
\title{Internal page for hidden aliases}
\usage{
groupFeatures(object, param, ...)
}
\value{
Not applicable
}
\description{
Internal page for hidden aliases
}
\examples{

NULL
}
\keyword{internal}
